+++ app/controllers/dashboard_controller.rb
@@ -3,8 +3,7 @@ class DashboardController < ApplicationController
   before_action :authenticate_user!
 
   def index
-    @created_rooms = current_user.created_rooms.order("created_at DESC")
-    @partipated_rooms = current_user.rooms.order("created_at DESC")
+    @user = UserPresenter.new(current_user)
   end
 
 end