+++ app/helpers/story_helper.rb
@@ -1,0 +1,11 @@
+module StoryHelper
+
+  def symbol_point_hash point
+    {
+      "coffee" => "☕",
+      "?"      => "⁉️",
+      "null"   => "skipped"
+    }[point] || point
+  end
+
+end