+++ db/migrate/20161209121853_add_stories_count_to_room.rb
@@ -1,0 +1,5 @@
+class AddStoriesCountToRoom < ActiveRecord::Migration[5.0]
+  def change
+    add_column :rooms, :stories_count, :integer
+  end
+end